requirements.txt

fastapi[standard]
sqlmodel
sqlalchemy
sqlite-utils

python-dateutil==2.8.2

requests
papermill

python-redmine
cryptography==42.0.8

papermill==2.3.4
nbconvert==6.5.0

pytest

ipython

# numpy
# pandas